Main mostly done

+++ ComplexPlane.h
@@ -27,7 +27,7 @@ class ComplexPlane
         void updateRender();
 
     private:
-        int countIterations(Vector2f coord);
+        size_t countIterations(Vector2f coord);
         void iterationsToRGB (size_t count, Uint8& r, Uint8& g, Uint8& b);
         Vector2f mapPixelToCoords(Vector2i mousePixel);
         VertexArray m_vArray;
